requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.10
PyMuPDF>=1.22.0

# Image Processing
Pillow>=10.0.0
easyocr>=1.7.0

# Excel
openpyxl>=3.1.0

# Web Requests
requests>=2.31.0

# Streamlit & Visualization
streamlit>=1.28.0
plotly>=5.0.0
matplotlib>=3.7.0

# Machine Learning & NLP
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0